Move 'random' command into extra_commands
